issue #191 Roth IRA implemented

+++ helpers/functions.py
@@ -852,7 +852,7 @@ def calculate_401k(
 def roth_ira(
     principal: float,
     interest_rate: float,
-    years: float,
+    years: int,
     tax_rate: float,
     annual_contribution: float,
 ):
